requirements.txt
python-telegram-bot==21.9
openpyxl==3.1.2